requirements.txt
requests>=2.31.0
python-telegram-bot>=20.0
flask>=2.3.0
websockets>=11.0
